TIP-872: Move missing classes
