+++ src/Akeneo/Channel/Bundle/Resources/config/routing.yml
@@ -14,3 +14,6 @@ external_api_channel:
     resource: '@AkeneoChannelBundle/Resources/config/routing/external_api/channel.yml'
     prefix: /api/rest/v1
 
+external_api_currency:
+    resource: '@AkeneoChannelBundle/Resources/config/routing/external_api/currency.yml'
+    prefix: /api/rest/v1
