+++ src/Akeneo/Channel/Bundle/Resources/config/routing/external_api/currency.yml
