+++ src/Akeneo/Pim/Enrichment/Bundle/Doctrine/ORM/Repository/ExternalApi/ProductRepository.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Akeneo\Tool\Bundle\ApiBundle\Doctrine\ORM\Repository;
+namespace Akeneo\Pim\Enrichment\Bundle\Doctrine\ORM\Repository\ExternalApi;
 
 use Akeneo\Pim\Enrichment\Component\Product\Repository\ExternalApi\ProductRepositoryInterface;
 use Doctrine\ORM\EntityManager;
