+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/repositories.yml
@@ -138,3 +138,10 @@ services:
         arguments: ['%pim_comment.entity.comment.class%']
         tags:
             - { name: 'pim_repository' }
+
+    pim_api.repository.product:
+        class: 'Akeneo\Pim\Enrichment\Bundle\Doctrine\ORM\Repository\ExternalApi\ProductRepository'
+        arguments:
+            - '@pim_catalog.object_manager.product'
+            - '%pim_catalog.entity.product.class%'
+            - '@pim_catalog.repository.product'
