+++ src/Akeneo/Tool/Bundle/ApiBundle/Resources/config/controllers.yml
@@ -2,7 +2,6 @@ parameters:
     pim_api.controller.token.class: Akeneo\Tool\Bundle\ApiBundle\Controller\TokenController
     pim_api.controller.root_endpoint.class: Akeneo\Tool\Bundle\ApiBundle\Controller\RootEndpointController
     pim_api.controller.media_file.class: Akeneo\Tool\Bundle\ApiBundle\Controller\MediaFileController
-    pim_api.controller.currency.class: Akeneo\Tool\Bundle\ApiBundle\Controller\CurrencyController
     pim_api.controller.measure_family.class: Akeneo\Tool\Bundle\ApiBundle\Controller\MeasureFamilyController
 
 services:
@@ -36,15 +35,6 @@ services:
             - '@pim_catalog.repository.product_model'
             - '@pim_api.updater.product_model'
             - '@pim_catalog.saver.product_model'
-            - '%pim_api.configuration%'
-
-    pim_api.controller.currency:
-        class: '%pim_api.controller.currency.class%'
-        arguments:
-            - '@pim_api.repository.currency'
-            - '@pim_serializer'
-            - '@pim_api.pagination.parameter_validator'
-            - '@pim_api.pagination.offset_hal_paginator'
             - '%pim_api.configuration%'
 
     pim_api.controller.measure_family:
