+++ src/Akeneo/Tool/Bundle/ApiBundle/Resources/config/routing.yml
@@ -2,10 +2,6 @@ media:
     resource: '@PimApiBundle/Resources/config/routing/media.yml'
     prefix: /rest/v1
 
-currency:
-    resource: '@PimApiBundle/Resources/config/routing/currency.yml'
-    prefix: /rest/v1
-
 fos_oauth_server_token:
     path: /oauth/v1/token
     defaults: { _controller: pim_api.controller.token:tokenAction }
